requirements.txt
fastapi
uvicorn
pygithub
requests
pyyaml
python-dotenv
aiohttp
pytest
httpx
pygithub
